mac/type: change array type from Any to Sendable
all values mapped to the array are immutable or value type values,
though the Any type implies any type including mutables or reference
types. all mpv_node data can be represented as the former.

change Any to Sendable, which implies mutable or values type values.

this is also a preparation for swift 6 and strict concurrency support.